+++ taxi-carpool/src/main/java/edu/kangwon/university/taxicarpool/party/PartyController.java
@@ -1,8 +1,8 @@
 package edu.kangwon.university.taxicarpool.party;
 
-import edu.kangwon.university.taxicarpool.party.dto.PartyDTO.PartyCreateRequestDTO;
-import edu.kangwon.university.taxicarpool.party.dto.PartyDTO.PartyResponseDTO;
-import edu.kangwon.university.taxicarpool.party.dto.PartyDTO.PartyUpdateRequestDTO;
+import edu.kangwon.university.taxicarpool.party.dto.PartyCreateRequestDTO;
+import edu.kangwon.university.taxicarpool.party.dto.PartyResponseDTO;
+import edu.kangwon.university.taxicarpool.party.dto.PartyUpdateRequestDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
