+++ taxi-carpool/src/main/java/edu/kangwon/university/taxicarpool/party/PartyService.java
@@ -7,9 +7,9 @@ import edu.kangwon.university.taxicarpool.chatting.MessageType;
 import edu.kangwon.university.taxicarpool.member.MemberEntity;
 import edu.kangwon.university.taxicarpool.member.MemberRepository;
 import edu.kangwon.university.taxicarpool.member.exception.MemberNotFoundException;
-import edu.kangwon.university.taxicarpool.party.dto.PartyDTO.PartyCreateRequestDTO;
-import edu.kangwon.university.taxicarpool.party.dto.PartyDTO.PartyResponseDTO;
-import edu.kangwon.university.taxicarpool.party.dto.PartyDTO.PartyUpdateRequestDTO;
+import edu.kangwon.university.taxicarpool.party.dto.PartyCreateRequestDTO;
+import edu.kangwon.university.taxicarpool.party.dto.PartyResponseDTO;
+import edu.kangwon.university.taxicarpool.party.dto.PartyUpdateRequestDTO;
 import edu.kangwon.university.taxicarpool.party.partyException.KakaoApiException;
 import edu.kangwon.university.taxicarpool.party.partyException.MemberAlreadyInPartyException;
 import edu.kangwon.university.taxicarpool.party.partyException.MemberNotInPartyException;
